[arc] Cache isARCInertBB in an analysis called ProgramTerminationAnalysis.
One important property that we take advantage of in ARC is that along
certain paths we know that the language leaks memory meaning that these
blocks are effectively inert from ARC's perspective.

In general these leaking blocks are when we know that the program is
going to terminate fatally through something along the lines of
fatalError.

This analysis just serves as a cache of this information since it is
starting to become used in multiple parts of the pass pipeline.