Revert "A small step towards removing C-style loops"
This reverts commit 0d001480a94f975355cd458973d5c51cef55bff1.

The commit that changed the iteration style from c-based loops into for-each
loops caused major regressions in our string benchmarks. Arnold believes that we
are making a different inlining decision in the for-each loops. We should
reapply this patch after we fix the optimizer

The regression was detected in rdar://23776732.