[SideEffectAnalysis] change the invalidation+update strategy.
Get rid of the UpdateSideEffects pass and recompute the side-effect information on demand.
If a function is invalidated also invalidate the side-effect information of a function's callers .
To limit compile time, the on-demand recomputation is only done for a limited call-graph depth.
This should not be a problem because the pass manager now iterates functions in bottom-up order.