[Driver] Properly handle new cascading dependencies.
The existing code was only recording a change in the /last/ dependency as needing
to rebuild the file. This could have caused things not to get rebuilt when they
should have.

In practice, a number of things have to coincide for this to cause problems:
- A dependency (let's say a type) has to be considered "modified" in this build.
- Some other file has to get rebuilt that didn't previously depend on the type,
  or only depended on it in a non-cascading way (e.g. it's only used in function
  bodies).
- The file had to not be rebuilt because of any cascading dependencies (or
  because it itself changed).
- The file has to now depend on the type in a cascading way, so that downstream
  files need to be rebuilt when they didn't before.

Noticed by inspection while fixing the previous issue.