[cmpcodesize] Move script into dir to package-ify
The cmpcodesize script is a little unwieldy--the same file contains
command-line argument parsing, otool invocation, and result
output formatting.

To split the script into several files, we'll need to turn it into a
Python package. In order to do so, we first move it into its own
directory.

Note that to invoke the script, you'll now need to run
`utils/cmpcodesize/cmpcodesize`. That additional path component is
inconvenient for now, but things will get better soon.