Clang importer: test that AnyObject lookup finds properties by name.
With the Swift name lookup tables enabled, we can find Objective-C
properties in Swift based on their property name, even when it differs
from the name of the getter. Verify this, which is
rdar://problem/14776565. The bug won't actually be considered fixed
until the Swift name lookup tables are enabled globally.