Add a condition on what the maximum of locations there are in a function for RLE to optimize
RLE is an iterative data flow. Functions with too many locations may take a long time for the
data flow to converge.

Once we move to a genset and killset for RLE. we should be able to lessen the condition a bit more.

I have observed no difference in # of redundant loads eliminated on the stdlib (currently we
eliminate 3862 redundant loads).