[Sema] Return nullptr less often when typechecking stmts. NFC.
There's a risk of untypechecked exprs like SequenceExpr leaking out to
the validator if we happen to generate a new statement but fail to
return it. For the sake of simplicity and consistency, try to always
return the statement that's been typechecked. The classes for which we
still return nullptr sometimes are the following: ForEachStmt,
ReturnStmt, BreakStmt, ContinueStmt, FallthroughStmt.