Runtime: Clean up logic for setting Objective-C name of generic classes
We're not currently doing it, but we will soon be able to use
swift_initializeSuperclass() for class layouts which are not
dependent on generic parameters. In this case, we still need
to set the Objective-C class name.

On the other hand, if we're doing resilient layout for a
non-generic class, we don't need to set the Objective-C class
name.

NFC since this isn't hooked up completely yet.