[lit] The cmp() function should be treated as gone in Python 3
As noted in the "What's New in Python 3.0" [1] documentation the Python
2 `cmp` function no longer exists in Python 3. The document goes on to
say that if you really need the `cmp` function you should define your
own.

Since the function appears to only be used once in the script I defined
a local function rather than shadow the Python 2 function. The local
function, `_cmp`, implements the recommended expression from the "What's
New in Python 3.0" document. The behavior is now defined on both Python
2 and 3.

[1] https://docs.python.org/3.0/whatsnew/3.0.html#ordering-comparisons