[Mangler Utils] Fix an off-by-one error in the calculation of the longest huffman encoding for our alphabet.
We calculate a conservative estimate of how long our bitstream needs to be to
hold the decoded string.  This commit fixes an off-by-one error in the
calculation of the longest possible encoding sequence for our alphabet.  The bug
was that we considered both the root of the tree and the leafs of the tree for
the distance calculation. However, only edges should be considered for the
length calculation because edges represent the one/zero bits.

This commit can improve the efficiency of our encoding because the APInts we
need to work with will be slightly shorter.