[Mangler] Shrink bitwidth after divmod while decoding fixed-width strings.
We represent strings encoded using fixed-width in alphabet that's not a power of two as:

a_0 * C^0 + a_1 * C^1 ... + a_n * C^n, where 'C' is the charset length and 'a'
is the encoded character..

To extract a character we need to do a 'modulo C' operation followed by a
'div C' operation for each character. This commit reduces the bit width of
the decoded number with each iteration to accelerate the divmod calculation.
With each iteration the APInt becomes shorter and the divmod operation becomes
faster.