[Compression] Increase the number of chars used on non APInt words.
This commit accelerates the performance of DecodeStringFromNumber by increasing
the number of characters that fit into a 64-bit word from 8 (conservative
estimate that we have 8bit ascii chars) to 10 (6-bit encoding).  We now compute
the number of characters that fit into a 64 bit word based on the length of the
character set at compile time.